pyproject.toml
[tool.poetry]
name = "ls"
version = "0.1.0"
description = ""
authors = ["mir21bek <mirbek.2121@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-jazzmin = "^2.6.0"
drf-yasg = "^1.21.7"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
celery = "^5.3.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.1.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
